tg_data_parser/requirements.txt
telethon
python-dotenv
loguru
uvicorn
fastapi
pymongo
elasticsearch